multiprocessing.atomic.AtomicReference is a multiprocessing Proxy
This makes the AtomicReference type a Proxy which allows its memory to
be shared between processes. This approach is a bit fragile, because it
requires running a multiprocessing.Manager which seems somewhat buggy.
Ideally multiprocessing would expose a way for memory to be shared
arbitrarily between processes, leaving safe management up to the
application. However there does not seem to be a mechanism for this so
we are left with this.